+++ src/demo_index.js
@@ -9,14 +9,21 @@ import {tad} from '@famiprog-foundation/tests-are-demo';
 import {ExpTestsAreDemo} from './testsAreDemo/ExpTestsAreDemo';
 import {DragToCreateTestsAreDemo} from './testsAreDemo/DragToCreateTestsAreDemo';
 import {TableTestsAreDemo} from './testsAreDemo/TableTestsAreDemo';
+import {SelectedItemsTestsAreDemo} from './testsAreDemo/SelectedItemsTestsAreDemo';
+import {ContextMenuTestsAreDemo} from './testsAreDemo/ContextMenuTestsAreDemo';
 
 ReactDOM.render(
   <TestsAreDemoAppWrapper
     importSemanticUiCss
     app={<DemoTimeline />}
     importTestsCallback={() => {
-      // tad.addTests(DragToCreateTestsAreDemo, ExpTestsAreDemo);
-      tad.addTests(TableTestsAreDemo);
+      tad.addTests(
+        DragToCreateTestsAreDemo,
+        ExpTestsAreDemo,
+        SelectedItemsTestsAreDemo,
+        ContextMenuTestsAreDemo,
+        TableTestsAreDemo
+      );
     }}
   />,
   document.getElementById('root')