+++ src/index.js
@@ -15,12 +15,11 @@ export * from './types';
 export {default as Timebar} from './components/timebar';
 export {default as TimelineBody} from './components/body';
 export {default as ItemRenderer} from './components/ItemRenderer';
-export {GroupRenderer} from './components/GroupRenderer';
-export {GroupHeaderRenderer as ColumnHeaderRenderer} from './components/GroupHeaderRenderer';
 export {Marker} from './components/Marker';
 export {default as Selectbox} from './components/selector';
 export {BackgroundLayer} from './components/BackgroundLayer';
 export {HighlightedInterval} from './components/HighlightedInterval';
+export {IAction, IActionParam, IActionParamForRun} from './components/ContextMenu/IAction';
 
 // consts
 export {timebarFormat} from './consts/timebarConsts';