+++ src/stories/itemRenderer/ItemRenderer.stories.tsx
@@ -10,11 +10,10 @@ import { Item } from '../../types';
 import { Table, Column, DataCell} from 'fixed-data-table-2';
 
 export default {
-  title: 'Features/Item Renderer',
-  component: Timeline
+  title: 'Features/Item Renderer'
 };
 
-export const Main: ComponentStory<typeof Timeline> = () => {
+export const Main = () => {
   const tasks: Item[] = [
     ...someTasks,
     {
@@ -94,7 +93,7 @@ Main.parameters = {
   ]
 };
 
-export const DefaultPropsForItemRenderer: ComponentStory<typeof Timeline> = () => {
+export const DefaultPropsForItemRenderer = () => {
   return (
     <>
       <Alert
@@ -135,7 +134,7 @@ type CustomTask = Item & {
   allTestsPassed?: boolean
 } 
 
-export const CustomItemRenderer: ComponentStory<typeof Timeline> = () => {
+export const CustomItemRenderer = () => {
   const tasks: CustomTask[] = [...someTasks];
   tasks[0].type = 'analysis';
   tasks[0].spentHours = 24;