+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## Unreleased
 
+* [In some places, key of type string was not supported](https://github.com/flower-platform/react-timeline-10000/pull/40). Now keys of type `number | string` are supported.
+
+## v2.2.0
+
 ### Added
 
 * [Click and double click handler for a group](https://github.com/flower-platform/react-timeline-10000/pull/11)
 * [Drag to create](https://github.com/flower-platform/react-timeline-10000/pull/25)
+* [Context menu and selection](https://github.com/flower-platform/react-timeline-10000/pull/37)
 
 ### Fixed